keywords parsers

+++ src/main/kotlin/lexicon/tokens/TokenTypes.kt
@@ -1,6 +1,6 @@
 package lexicon.tokens
 
-enum class TokenTypes (val flag: Int){
+enum class TokenTypes (val flag: Byte){
     COMMENT                 (0b00000001),
     IDENTIFIER              (0b00000010),
     KEYWORD                 (0b00000100),
@@ -8,5 +8,5 @@ enum class TokenTypes (val flag: Int){
     REAL_LITERAL            (0b00010000),
     CHARACTER_LITERAL       (0b00100000),
     STRING_LITERAL          (0b01000000),
-    OPERATOR_OR_PUNCTUATOR  (0b10000000)
+    OPERATOR_OR_PUNCTUATOR  (0b10000000.toByte())
 }
